Fixes dynamic head tags example
While a key `frontMatter` does exist inside the object returned from `useConfig()`, the object is empty. The rest of the values we expect further in the example (like `frontMatter.title` and `frontMatter.description`) are present in the object returned from `useConfig()`. So we can do two things to improve this:

1. Not destructure and have `frontMatter` represent the values returned from `useConfig()` or

2. Destructure the values directly like so: `const { title, description } = useConfig()`