feat(theme): Content width css variable for easier customization
This moves the content width that was used throughout nextra
`max-w-[90rem]` to a css variable. This way the changing the value in
the future should be easier, but it also makes it possible to customize
the max width of the content for people using nextra without needing to
overload a tailwind class in a hacky way.